Broken Heliolisk Crest
It's supposed to change the Normal-type to Water/Rock/Fire depending on its ability.. now it does this but it also just changes Heliolisk's in-battle ability to Solar Power.
It's *also* supposed to change its normal moves to that type with 1.2x power, but it just doesn't for whatever reason. ;-;